# Improve SmallRng initialization performance