- [x] Added a `CHANGELOG.md` entry

# Summary

Improve initialization speed for SmallRng

# Motivation

`SmallRng` is plenty fast, but the initialization speed was suboptimal.

# Details

An example binary was added to check for code generation with `cargo asm`.

Benchmarks were also added. The results are shown below:

```
before
test init_from_seed_smallrng  ... bench:           0.84 ns/iter (+/- 0.06)
test init_from_u64_smallrng  ... bench:           6.19 ns/iter (+/- 0.58)
test init_smallrng           ... bench:           9.45 ns/iter (+/- 0.18)

after
test init_from_seed_smallrng  ... bench:           0.85 ns/iter (+/- 0.08)
test init_from_u64_smallrng  ... bench:           3.42 ns/iter (+/- 0.09)
test init_smallrng           ... bench:           6.65 ns/iter (+/- 0.15)
```