@dhardy 2024-09-10
Reminder: changes are needed to ensure the API remains consistent on 32-bit and 64-bit platforms.

Also, these new benchmarks can't be merged since #1490 moved all benchmarks to use Criterion.